refactor(meta): enable periodic full GC
