+++ src/config/example.toml
@@ -7,7 +7,8 @@ metrics_level = 0
 telemetry_enabled = true
 
 [meta]
-min_sst_retention_time_sec = 604800
+min_sst_retention_time_sec = 86400
+full_gc_interval_sec = 86400
 collect_gc_watermark_spin_interval_sec = 5
 periodic_compaction_interval_sec = 60
 vacuum_interval_sec = 30
